# interface-service/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Interface"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/reminder/create": {
      "post": {
        "operationId": "Interface_ReminderCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ReminderCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReminderCreateRequest"
            }
          }
        ],
        "tags": [
          "Interface"
        ]
      }
    },
    "/reminder/remove": {
      "delete": {
        "operationId": "Interface_ReminderRemove",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ReminderRemoveResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReminderRemoveRequest"
            }
          }
        ],
        "tags": [
          "Interface"
        ]
      }
    },
    "/reminder/update": {
      "put": {
        "operationId": "Interface_ReminderUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ReminderUpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReminderUpdateRequest"
            }
          }
        ],
        "tags": [
          "Interface"
        ]
      }
    },
    "/reminder/{id}/get": {
      "get": {
        "operationId": "Interface_ReminderGet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ReminderGetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Interface"
        ]
      }
    },
    "/reminders": {
      "get": {
        "operationId": "Interface_ReminderList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ReminderListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Interface"
        ]
      }
    }
  },
  "definitions": {
    "Reminder": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date-time"
        },
        "text": {
          "type": "string"
        },
        "id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "ReminderCreateRequest": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date-time"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "ReminderCreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "ReminderGetResponse": {
      "type": "object",
      "properties": {
        "reminder": {
          "$ref": "#/definitions/Reminder"
        }
      }
    },
    "ReminderListResponse": {
      "type": "object",
      "properties": {
        "reminders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Reminder"
          }
        }
      }
    },
    "ReminderRemoveRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "ReminderRemoveResponse": {
      "type": "object"
    },
    "ReminderUpdateRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "ReminderUpdateResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}